% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_product.R
\name{find_product}
\alias{find_product}
\title{One line title}
\usage{
find_product(data, target)
}
\arguments{
\item{data}{Argument description.}

\item{target}{Arguments to be passed to the model specific MASEM function.}
}
\value{
Specify what are returned.
}
\description{
One line description
}
\details{
Details
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link{functionname}}
}
