% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_mod_sem}
\alias{sample_mod_sem}
\title{Sample dataset for testing}
\format{
An object of class \code{data.frame} with 500 rows and 6 columns.
}
\usage{
sample_mod_sem
}
\description{
A data set with six variables, x1, x2, m1, m2, y.
}
\details{
Fit the following model:

#mod <-
#"
#m1 ~ x1 + x2
#m2 ~ x1 + x2 + x1x2
#y1 ~ m1 + x1
#y2 ~ m1 + m2
#"
}
\keyword{datasets}
