% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmodsem-package.R
\docType{package}
\name{semlbci-package}
\alias{semlbci-package}
\title{plotmodsem: Plotting moderation effects in path models}
\description{
Plot the selected moderation effect in a path model fitted by an
'             structural equation modeling package (e.g., lavaan)
}
\seealso{
\code{\link[=plotmod]{plotmod()}}
}
