% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmod_new.R
\name{plotmod}
\alias{plotmod}
\title{Plot the moderation effect in
a path model (WIP: Not Ready)}
\usage{
plotmod(
  fit,
  x,
  y,
  wlevels,
  w_type = "auto",
  w_method = "sd",
  sd_from_mean = NULL,
  percentiles = NULL,
  mod_levels_list_args = list(),
  standardized_x = FALSE,
  standardized_y = FALSE,
  x_label,
  w_label = "Moderator(s)",
  y_label,
  title,
  x_from_mean_in_sd = 1,
  x_method = c("sd", "percentile"),
  x_percentiles = c(0.16, 0.84),
  x_sd_to_percentiles = NA,
  note_standardized = TRUE,
  no_title = FALSE,
  line_width = 1,
  point_size = 5,
  graph_type = c("default", "tumble"),
  ...
)
}
\arguments{
\item{fit}{The fit object. Can be a
\link[lavaan:lavaan-class]{lavaan::lavaan} object or a
list of \code{\link[=lm]{lm()}} outputs.
It can also be
a \code{lavaan.mi} object
returned by
\code{\link[semTools:runMI]{semTools::runMI()}} or
its wrapper, such as \code{\link[semTools:runMI]{semTools::sem.mi()}}.}

\item{x}{Character. The name of the
predictor at the start of the path.}

\item{y}{Character. The name of the
outcome variable at the end of the
path.}

\item{wlevels}{The output of
\code{\link[manymome:merge_mod_levels]{manymome::merge_mod_levels()}}, or the
moderator(s) to be passed to
\code{\link[manymome:mod_levels]{manymome::mod_levels_list()}}. If all the
moderators can be represented by one
variable, that is, each moderator is
(a) a numeric variable, (b) a
dichotomous categorical variable, or
(c) a factor or string variable used
in \code{\link[=lm]{lm()}} in \code{fit}, then it is a
vector of the names of the moderators
as appeared in the data frame. If at
least one of the moderators is a
categorical variable represented by
more than one variable, such as
user-created dummy variables used in
\code{\link[lavaan:sem]{lavaan::sem()}}, then it must be a
list of the names of the moderators,
with such moderators represented by a
vector of names. For example:
\verb{list("w1", c("gpgp2", "gpgp3")}, the
first moderator \code{w1} and the second
moderator a three-categorical
variable represented by \code{gpgp2} and
\code{gpgp3}.}

\item{w_type}{Character. Whether the
moderator is a \code{"numeric"} variable
or a \code{"categorical"} variable. If
\code{"auto"}, the function will try to
determine the type automatically.
See \code{\link[=mod_levels_list]{mod_levels_list()}} for further
information.}

\item{w_method}{Character, either
\code{"sd"} or \code{"percentile"}. If \code{"sd"},
the levels are defined by the
distance from the mean in terms of
standard deviation. if
\code{"percentile"}, the levels are
defined in percentiles.  See
\code{\link[=mod_levels_list]{mod_levels_list()}} for further
information.}

\item{sd_from_mean}{A numeric vector.
Specify the distance in standard
deviation from the mean for each
level. Default is \code{c(-1, 0, 1)} when
there is only one moderator, and
\code{c(-1, 1)} when there are more than
one moderator. Ignored if \code{w_method}
is not equal to \code{"sd"}. See
\code{\link[=mod_levels_list]{mod_levels_list()}} for further
information.}

\item{percentiles}{A numeric vector.
Specify the percentile (in
proportion) for each level. Default
is \code{c(.16, .50, .84)} if there is one
moderator, and \code{c(.16, .84)} when
there are more than one moderator.
Ignored if \code{w_method} is not equal to
\code{"percentile"}. See
\code{\link[=mod_levels_list]{mod_levels_list()}} for further
information.}

\item{mod_levels_list_args}{Additional arguments to be passed to
\code{\link[=mod_levels_list]{mod_levels_list()}} if it is called
for creating the levels of
moderators. Default is \code{list()}.}

\item{standardized_x}{Logical.
Whether \code{x} will be standardized.
Default is \code{FALSE}.}

\item{standardized_y}{Logical.
Whether \code{y} will be standardized.
Default is \code{FALSE}.}

\item{x_label}{The label for the
X-axis. Default is the value of the
predictor in the output of
\code{\link[manymome:cond_indirect]{manymome::cond_indirect_effects()}}.}

\item{w_label}{The label for the
legend for the lines. Default is
\code{"Moderator(s)"}.}

\item{y_label}{The label for the
Y-axis. Default is the name of the
response variable in the model.}

\item{title}{The title of the graph.
If not supplied, it will be generated
from the variable names or labels (in
\code{x_label}, \code{y_label}, and \code{w_label}).
If \code{""}, no title will be printed.
This can be used when the plot is for
manuscript submission and figures are
required to have no titles.}

\item{x_from_mean_in_sd}{How many SD
from mean is used to define "low" and
"high" for the focal variable.
Default is 1.}

\item{x_method}{How to define "high"
and "low" for the focal variable
levels. Default is in terms of the
standard deviation of the focal
variable, \code{"sd"}. If equal to
\code{"percentile"}, then the percentiles
of the focal variable in the dataset
is used.}

\item{x_percentiles}{If \code{x_method} is
\code{"percentile"}, then this argument
specifies the two percentiles to be
used, divided by 100. It must be a
vector of two numbers. The default
is \code{c(.16, .84)}, the 16th and 84th
percentiles, which corresponds
approximately to one SD below and
above mean for a normal distribution,
respectively.}

\item{x_sd_to_percentiles}{If
\code{x_method} is \code{"percentile"} and this
argument is set to a number, this
number will be used to determine the
percentiles to be used. The lower
percentile is the percentile in a
normal distribution that is
\code{x_sd_to_percentiles} SD below the
mean. The upper percentile is the
percentile in a normal distribution
that is \code{x_sd_to_percentiles} SD
above the mean. Therefore, if
\code{x_sd_to_percentiles} is set to 1,
then the lower and upper percentiles
are 16th and 84th, respectively.
Default is \code{NA}.}

\item{note_standardized}{If \code{TRUE},
will check whether a variable has SD
nearly equal to one. If yes, will
report this in the plot. Default is
\code{TRUE}.}

\item{no_title}{If \code{TRUE}, title will
be suppressed. Default is \code{FALSE}.}

\item{line_width}{The width of the
lines as used in
\code{ggplot2::geom_segment()}. Default is
1.}

\item{point_size}{The size of the
points as used in
\code{ggplot2::geom_point()}. Default is
5.}

\item{graph_type}{If \code{"default"}, the
typical line-graph with equal
end-points will be plotted. If
\code{"tumble"}, then the tumble graph
proposed by Bodner (2016) will be
plotted. Default is \code{"default"}.}

\item{...}{Arguments to be passed to
the plot method of the output of
\code{\link[manymome:cond_indirect]{manymome::cond_indirect_effects()}}.}
}
\value{
A \code{ggplot2} graph.
}
\description{
Plot the moderation
effect in a path model
}
\details{
This function extracts the
information stored in the \code{lavaan}
fit object to plot a line graph,
one for the relation between the
focal variable (\code{x}) and the outcome
variable (\code{y}) for each level of
the moderator
(\code{w}).

It calls several functions from
the package \code{manymome} internally.
}
\examples{
\dontrun{
# To be prepared
}
}
