% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmod.R
\name{plotmod}
\alias{plotmod}
\title{Plot the moderation effect in a path model}
\usage{
plotmod(
  fit,
  y,
  x,
  w,
  xw,
  x_label,
  w_label,
  y_label,
  title,
  a_shift = 0,
  expansion = 0.1,
  standardized = FALSE,
  digits = 3
)
}
\arguments{
\item{fit}{The output from \link[lavaan:lavaan]{lavaan::lavaan} or its wrapper, such as
\link[lavaan:sem]{lavaan::sem}.}

\item{y}{Character. The name of the outcome variable as in the data set in \code{fit}.}

\item{x}{Character. The name of the focal variable as in the data set in \code{fit}.}

\item{w}{Character. The name of the moderator as in the data set in \code{fit}.}

\item{xw}{Charcter. The name of the product term, \code{x * w}. If not supplied,
The function will try to find it in the data set.}

\item{x_label}{The label for the X-axis. Default is the vlaues of \code{x}.}

\item{w_label}{The label for the legend for the lines. Default is the value of\code{w}.}

\item{y_label}{The label for the Y-axis. Default is the value of \code{y}.}

\item{title}{The title of the graph. If not supplied, will be generated from the variable
names.}

\item{a_shift}{Default is 0. Can be ignored for now.}

\item{expansion}{How much tha lower and upper limits of the axis will be adjusted.}

\item{standardized}{Logical. Plot the moderation effect in standardized metric. All three
variables, \code{x}, \code{w}, and \code{y} will be standardized. Default
is \code{FALSE}}

\item{digits}{Number of decimal digits to print. Default is 3.}
}
\value{
A \link{ggplot2} graph.
}
\description{
Plot the moderation effect in a path model
}
\details{
This function extracts the information stored
in the \code{lavaan} fit object to plot a two-line
graph, one for the relation between the focal variale (\code{x}) and
the outcome variable (\code{y}) when the moderator (\code{w}) is one stanard
deviation below mean, and one when the moderator is one standard
deviation above mean.
}
\examples{
\dontrun{
# To be prepared
}
}
