% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_mod_sem}
\alias{sample_mod_sem}
\title{Sample dataset for testing}
\format{
A data frame with 300 rows
and 5 variables:
\describe{
\item{x1}{Predictor. Numeric.}
\item{x2}{Moderator. Numeric}
\item{m1}{Mediator. Numeric.}
\item{m2}{Mediator. Numeric.}
\item{y1}{Outcome variable. Numeric.}
\item{y2}{Outcome variable. Numeric.}
}
}
\usage{
sample_mod_sem
}
\description{
A data set with 500 cases and six variables, x1, x2, m1, m2, y.
}
\examples{
library(lavaan)
library(ggplot2)
dat <- sample_mod_sem
dat$x1x2 <- dat$x1 * dat$x2
mod <-
"
m1 ~ x1 + x2
m2 ~ bx1 * x1 + bx2 * x2 + bx1x2 * x1x2
y1 ~ m1 + x1
y2 ~ m1 + m2
x2 ~ x2mean * 1
m2 ~ m2intercept * 1
x2 ~~ x2v * x2
bx1_lo := bx1 + bx1x2 * (x2mean - sqrt(x2v))
bx1_hi := bx1 + bx1x2 * (x2mean + sqrt(x2v))
b0_lo := m2intercept + bx2 * (x2mean - sqrt(x2v))
b0_hi := m2intercept + bx2 * (x2mean + sqrt(x2v))
"
fit <- sem(model = mod, data = dat, meanstructure = TRUE, fixed.x = FALSE)
parameterEstimates(fit)
p <- plotmod(fit, y = "m2", x = "x1", w = "x2", xw = "x1x2",
             a_shift = coef(fit)["m2intercept"])
p
}
\keyword{datasets}
